Let’s look at each column in detail, which will help us understand the range, unique values, and data types of both numeric and categorical columns.
Column 'Name' has 13838 unique values and is of type object.
Column 'Year' has 102 unique values and is of type object.
Column 'Duration' has 182 unique values and is of type object.
Column 'Genre' has 485 unique values and is of type object.
Column 'Rating' has 84 unique values and is of type float64.
Column 'Votes' has 2034 unique values and is of type object.
Column 'Director' has 5938 unique values and is of type object.
Column 'Actor 1' has 4718 unique values and is of type object.
Column 'Actor 2' has 4891 unique values and is of type object.
Column 'Actor 3' has 4820 unique values and is of type object.

Co-Authored-By: Sumit Agnihotri <163757254+Sumit-Agnihotri@users.noreply.github.com>